+++ src/components/table/index.vue
@@ -1,20 +1,35 @@
 <template>
     <div style="position: relative">
         <div v-if="columnsFixed.length > 0" :style="{width: columnsFixedWidth+'px'}" style="position: absolute; top: 0;z-index: 1;overflow: hidden">
-            <table-body :columns="columnsFixed" :data-source="dataSource" :bordered = "bordered" :size="size"/>
+            <table-body
+                :on-change="handleTableChange"
+                :columns="columnsFixed"
+                :data-source="dataSource"
+                :bordered = "bordered"
+                :stripe = "stripe"
+                :size="size"
+                :sort-field="sortField"
+                :sort-type="sortType"
+            />
         </div>
-        <div style="overflow: auto">
-            <table-body :columns="columns" :data-source="dataSource" :bordered = "bordered" :size="size"/>
+        <div class="dd-table-container">
+            <table-body
+                :on-change="handleTableChange"
+                :columns="columns"
+                :data-source="dataSource"
+                :bordered = "bordered"
+                :stripe = "stripe"
+                :size="size"
+                :sort-field="sortField"
+                :sort-type="sortType"
+            />
         </div>
     </div>
 </template>
 <style>
-    table {
-        background: #fff;
-        border-collapse: collapse;
-        border-spacing: 0;
-        table-layout: fixed;
-        width: 100%;
+    .dd-table-container {
+        box-shadow: 0 0 5px 0 rgba(0,0,0,0.15);
+        overflow: auto;
     }
 </style>
 <script>
@@ -30,11 +45,17 @@
             size: {
                 type: String,
                 default: 'default'
-            }
+            },
+            stripe: {
+                type: Boolean,
+                default: false
+            },
+            onChange: Function
         },
         data() {
             return {
-               
+                sortField: undefined,
+                sortType: undefined
             }
         },
         computed: {
@@ -56,6 +77,11 @@
         methods: {
             expand(row) {
                 this.$set(row, 'expand', !row.expand);
+            },
+            handleTableChange(data) {
+                this.sortField = data.sortField;
+                this.sortType = data.sortType;
+                this.onChange && this.onChange(data);
             }
         },
         components: {